refactor(dbt): rename `resources_v2` to `resource`
