--- python_modules/libraries/dagster-dbt/dagster_dbt/core/__init__.py
@@ -1,5 +1,0 @@
-from .resources_v2 import (
-    DbtCliEventMessage as DbtCliEventMessage,
-    DbtCliInvocation as DbtCliInvocation,
-    DbtCliResource as DbtCliResource,
-)
