+++ python_modules/libraries/dagster-dbt/dagster_dbt_tests/core/test_def_metadata.py
@@ -5,7 +5,7 @@ import pytest
 from dagster import AssetExecutionContext
 from dagster._core.definitions.metadata.metadata_set import TableMetadataSet
 from dagster_dbt.asset_decorator import dbt_assets
-from dagster_dbt.core.resources_v2 import DbtCliResource
+from dagster_dbt.core.resource import DbtCliResource
 
 pytestmark: pytest.MarkDecorator = pytest.mark.derived_metadata
